refactor: move selection content to right component
