+++ front/components/agent_builder/skills/skillSheet/hooks.ts
@@ -1,0 +1,79 @@
+import { useCallback, useMemo, useState } from "react";
+
+import { useAgentBuilderContext } from "@app/components/agent_builder/AgentBuilderContext";
+import type { AgentBuilderSkillsType } from "@app/components/agent_builder/AgentBuilderFormContext";
+import type { SelectionMode } from "@app/components/agent_builder/skills/skillSheet/types";
+import { useSkillConfigurations } from "@app/lib/swr/skill_configurations";
+import type { SkillConfigurationType } from "@app/types/assistant/skill_configuration";
+
+export const useLocalSelectedSkills = ({
+  mode,
+  onSave,
+  onClose,
+}: {
+  mode: SelectionMode;
+  onSave: (skills: AgentBuilderSkillsType[]) => void;
+  onClose: () => void;
+}) => {
+  const { owner } = useAgentBuilderContext();
+  const [searchQuery, setSearchQuery] = useState("");
+  const [localSelectedSkills, setLocalSelectedSkills] = useState<
+    AgentBuilderSkillsType[]
+  >(mode.selectedSkills);
+
+  const { skillConfigurations, isSkillConfigurationsLoading } =
+    useSkillConfigurations({
+      owner,
+      disabled: !open,
+    });
+
+  const selectedSkillIds = useMemo(
+    () => new Set(localSelectedSkills.map((s) => s.sId)),
+    [localSelectedSkills]
+  );
+
+  const filteredSkills = useMemo(() => {
+    if (!searchQuery.trim()) {
+      return skillConfigurations;
+    }
+    const query = searchQuery.toLowerCase();
+    return skillConfigurations.filter(
+      (skill) =>
+        skill.name.toLowerCase().includes(query) ||
+        skill.description.toLowerCase().includes(query)
+    );
+  }, [skillConfigurations, searchQuery]);
+
+  const handleSkillToggle = (skill: SkillConfigurationType) => {
+    setLocalSelectedSkills((prev) => {
+      const isAlreadySelected = prev.some((s) => s.sId === skill.sId);
+      if (isAlreadySelected) {
+        return prev.filter((s) => s.sId !== skill.sId);
+      } else {
+        return [
+          ...prev,
+          {
+            sId: skill.sId,
+            name: skill.name,
+            description: skill.description,
+          },
+        ];
+      }
+    });
+  };
+
+  const handleSave = useCallback(() => {
+    onSave(localSelectedSkills);
+    onClose();
+  }, [localSelectedSkills, onSave, onClose]);
+
+  return {
+    handleSkillToggle,
+    handleSave,
+    filteredSkills,
+    isSkillConfigurationsLoading,
+    searchQuery,
+    selectedSkillIds,
+    setSearchQuery,
+  };
+};
