+++ front/components/agent_builder/skills/skillSheet/types.tsx
@@ -1,3 +1,4 @@
+import type { AgentBuilderSkillsType } from "@app/components/agent_builder/AgentBuilderFormContext";
 import type { SkillConfigurationType } from "@app/types/assistant/skill_configuration";
 
 export const SKILLS_SHEET_PAGE_IDS = {
@@ -6,22 +7,24 @@ export const SKILLS_SHEET_PAGE_IDS = {
 } as const;
 
 export type SkillsSheetMode =
-  | { type: typeof SKILLS_SHEET_PAGE_IDS.SELECTION }
+  | {
+      type: typeof SKILLS_SHEET_PAGE_IDS.SELECTION;
+      selectedSkills: AgentBuilderSkillsType[];
+    }
   | {
       type: typeof SKILLS_SHEET_PAGE_IDS.INFO;
       skillConfiguration: SkillConfigurationType;
-      source: "toolDetails" | "addedTool";
+      previousMode: SelectionMode;
     };
+
+export type SelectionMode = Extract<
+  SkillsSheetMode,
+  { type: typeof SKILLS_SHEET_PAGE_IDS.SELECTION }
+>;
 
 export type PageContentProps = {
   mode: SkillsSheetMode;
-  searchQuery: string;
-  setSearchQuery: (query: string) => void;
-  isSkillConfigurationsLoading: boolean;
-  filteredSkills: SkillConfigurationType[];
-  selectedSkillIds: Set<string>;
-  handleSkillToggle: (skill: SkillConfigurationType) => void;
   onModeChange: (mode: SkillsSheetMode | null) => void;
   onClose: () => void;
-  handleSave: () => void;
+  onSave: (skills: AgentBuilderSkillsType[]) => void;
 };
