+++ front/components/agent_builder/skills/skillSheet/utils.tsx
@@ -1,7 +1,8 @@
 import type { ButtonProps, MultiPageSheetPage } from "@dust-tt/sparkle";
 import React from "react";
 
-import { SelectionPageContent } from "@app/components/agent_builder/skills/skillSheet/SelectionContent";
+import { useLocalSelectedSkills } from "@app/components/agent_builder/skills/skillSheet/hooks";
+import { SelectionPageContent } from "@app/components/agent_builder/skills/skillSheet/SelectionPage";
 import type { PageContentProps } from "@app/components/agent_builder/skills/skillSheet/types";
 import { SKILLS_SHEET_PAGE_IDS } from "@app/components/agent_builder/skills/skillSheet/types";
 import { assertNever } from "@app/types";
@@ -11,16 +12,28 @@ export function getPageAndFooter(props: PageContentProps): {
   leftButton?: ButtonProps & React.RefAttributes<HTMLButtonElement>;
   rightButton?: ButtonProps & React.RefAttributes<HTMLButtonElement>;
 } {
-  switch (props.mode.type) {
+  const mode = props.mode;
+  switch (mode.type) {
     case SKILLS_SHEET_PAGE_IDS.SELECTION:
+      // eslint-disable-next-line react-hooks/rules-of-hooks
+      const { handleSave } = useLocalSelectedSkills({
+        mode,
+        onSave: props.onSave,
+        onClose: props.onClose,
+      });
+
       return {
         page: {
           title: "Add skills",
-          id: props.mode.type,
-          content: <SelectionPageContent {...props} mode={props.mode} />,
+          id: mode.type,
+          content: <SelectionPageContent {...props} mode={mode} />,
         },
         leftButton: getCancelButton(props.onClose),
-        rightButton: getSaveButton(props.handleSave),
+        rightButton: {
+          label: "Add skills",
+          onClick: handleSave,
+          variant: "primary",
+        },
       };
     case SKILLS_SHEET_PAGE_IDS.INFO:
       return {
@@ -32,14 +45,11 @@ export function getPageAndFooter(props: PageContentProps): {
         leftButton: {
           label: "Back",
           variant: "outline",
-          onClick: () =>
-            props.onModeChange({
-              type: SKILLS_SHEET_PAGE_IDS.SELECTION,
-            }),
+          onClick: () => props.onModeChange(mode.previousMode),
         },
       };
     default:
-      assertNever(props.mode);
+      assertNever(mode);
   }
 }
 
@@ -47,10 +57,4 @@ export const getCancelButton = (onClose: () => void) => ({
   label: "Cancel",
   variant: "outline",
   onClick: onClose,
-});
-
-export const getSaveButton = (handleSave: () => void) => ({
-  label: "Save",
-  variant: "primary",
-  onClick: handleSave,
 });
